+++ apps/api/src/app.module.ts
@@ -11,6 +11,7 @@ import { AuctionsModule } from './auctions/auctions.module';
 import { ScheduleModule } from '@nestjs/schedule';
 import { BiddingModule } from './bidding/bidding.module';
 import { WatchlistsModule } from './watchlists/watchlists.module';
+import { NotificationsModule } from './notifications/notifications.module';
 
 @Module({
   imports: [
@@ -26,6 +27,7 @@ import { WatchlistsModule } from './watchlists/watchlists.module';
     AuctionsModule,
     BiddingModule,
     WatchlistsModule,
+    NotificationsModule,
   ],
   controllers: [AppController],
   providers: [AppService],